move utils into separate file
